Webhook handler header check as middleware
